refectored ApplicationModel.kt
